\name{bridge.dat}
\alias{bridge.dat}
\docType{data}

\title{Bridge data}

\description{
A modification of the 60:80 data to reduce the sharpness of the division between groups. 
}

\usage{data(bridge.dat)}

\format{
  A data frame with 170 observations on the following 2 variables.
  \describe{
    \item{y1}{a numeric vector.}
    \item{y2}{a numeric vector.}
  }
}

\details{
In the 60:80 data and in the "three clusters and two outliers" data, the groups of observations were clearly separated.  However such complete separation is rare in real world examples where clusters often overlap.  This data is a modification of the 60:80 data to reduce the sharpness of the division between groups.
}

\source{
Atkinson Riani and Cerioli (2004), p. 590-591; http://www.riani.it/arc.
}

\keyword{datasets}

