\name{iris.dat}
\alias{iris.dat}
\docType{data}

\title{Iris data}

\description{
The data contain measurements on three species of iris.  The species are: 1) Iris setosa, 2) Iris versicolor, and 3) Iris virginica.  Four measurements of characteristic dimensions of the flowers were made on fifty flowers from each species.
}

\usage{data(iris.dat)}

\format{
  A data frame with 150 observations on the following 4 variables.
  \describe{
    \item{y1}{a numeric vector, the sepal length.}
    \item{y2}{a numeric vector, the sepal width.}
    \item{y3}{a numeric vector, the petal length.}
    \item{y4}{a numeric vector, the petal width.}
  }
}

\source{
The data were published by Anderson (1935) from measurements taken on plants in the Gaspe Peninsula, Quebec.  The three species are blue-flowered water loving irises, or flags, similar to the European yellow flag. The data were analysed by Fisher (1936) as an example of discriminant analysis and are often known as "Fisher's Iris data".
}

\references{
Atkinson, Riani and Cerioli (2004), p. 576-578; http://www.riani.it/arc.
}

\examples{
data(iris.dat)
}

\keyword{datasets}
