\name{record.dat}
\alias{record.dat}
\docType{data}

\title{Track rercords for women}

\description{
This data set contains women's athletic records for 55 countries.
}

\usage{data(record.dat)}

\format{
  A data frame with 55 observations on the following 7 variables.
  \describe{
    \item{y1}{a numeric vector, 100 metre record in seconds.}
    \item{y2}{a numeric vector, 200 metre record in seconds.}
    \item{y3}{a numeric vector, 400 metre record in seconds.}
    \item{y4}{a numeric vector, 800 metre record in minutes.}
    \item{y5}{a numeric vector, 1500 metre record in minutes.}
    \item{y6}{a numeric vector, 3000 metre record in minutes.}
    \item{y7}{a numeric vector, marathon record in minutes.}
  }
}

\details{
These data are taken from a handbook prepared for the 1984 Olympic games in Los Angeles.  They therefore come from an interesting period in the history of women's athletics, when there were, now authenticated, allegations about the treatment of female athletes, especially in communist countries, with male sex hormones.
}

\source{
Johnson and Wichern (1997), pp. 44-45.
}

\references{
Atkinson, Riani and Cerioli (2004), p. 558-559; http://www.riani.it/arc.
}

\examples{
data(record.dat)
}

\keyword{datasets}
